+++ composer.json
@@ -40,7 +40,7 @@
     },
     "autoload": {
         "psr-4": {
-            "rollun\\installer\\": "src/"
+            "rollun\\installer\\": "src/Installer/src"
         }
     },
     "bin": [
@@ -50,6 +50,9 @@
         "psr-4": {
             "rollun\\test\\installer\\": "tests/src"
         }
+    },
+    "extra": {
+      "class": "rollun\\installer\\InstallerPlugin"
     },
     "scripts": {
         "post-create-project-cmd": [
